use named route as adress for widgets and define
